requirements.txt
flask
flask-cors
openai
anthropic
requests
bcrypt
pyjwt
google-generativeai>=0.3.0
llama-cpp-python>=0.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
flask>=2.3.0
